common assertion files
