+++ envoy-control-tests/src/main/kotlin/pl/allegro/tech/servicemesh/envoycontrol/assertions/ResponseAssertions.kt
@@ -1,0 +1,9 @@
+package pl.allegro.tech.servicemesh.envoycontrol.assertions
+
+import okhttp3.Response
+import org.assertj.core.api.ObjectAssert
+
+fun ObjectAssert<Response>.isOk(): ObjectAssert<Response> {
+    matches { it.isSuccessful }
+    return this
+}
